Fix for slowdown when playing sounds quickly.
Up until now, ObjectAL has monitored every OpenAL call for an error, and would reset the audio session when this happened in order to get around playback bugs in iOS that would cause OpenAL to get stuck in a bad state and stop playing audio entirely (resetting the audio session would fix it).

Unfortunately, there is another OpenAL issue where it will start to intermittently return a -1 error code (which is itself an invalid error code) if you try to play buffers too often. This would cause it to output a log message and reset the audio session, an expensive operation. At sufficiently high playback rates, it could grind the app to a halt.

To fix it, I've disabled the audio reset when a -1 error code is returned. This gets around the issue, but of course now it will never reset the audio on a -1 code. Since the "bad state" bug was in iOS 4, it may not even be an issue anymore, but if it is still an issue, this fix may expose that old bug once again, disabling audio entirely in certain rare situations.

I've also added another demo that allows you to test fast buffer play rates.